Revert "rust fmt"
This reverts commit db604eaf5241acea442d29fa3d38830fae8b9617.

revert

+++ src/build/mod.rs
@@ -3,20 +3,16 @@ use std::path::Path;
 use std::process::{Command, Stdio};
 
 use reqwest;
-use serde::{Deserialize, Serialize};
+use serde::{Serialize, Deserialize};
 use tokio::fs;
 
-use super::setup::{
-    check_js_deps, check_py_deps, check_rust_deps, get_deps, set_newest_valid_node_version,
-    OldNodeVersion, REQUIRED_PY_PACKAGE,
-};
+use super::setup::{check_js_deps, check_py_deps, check_rust_deps, get_deps, set_newest_valid_node_version, OldNodeVersion, REQUIRED_PY_PACKAGE};
 
 const PY_VENV_NAME: &str = "process_env";
 const JAVASCRIPT_SRC_PATH: &str = "src/lib.js";
 const PYTHON_SRC_PATH: &str = "src/lib.py";
 const RUST_SRC_PATH: &str = "src/lib.rs";
-const KINODE_WIT_URL: &str =
-    "https://raw.githubusercontent.com/kinode-dao/kinode-wit/v0.7.0-alpha/kinode.wit";
+const KINODE_WIT_URL: &str = "https://raw.githubusercontent.com/kinode-dao/kinode-wit/v0.7.0-alpha/kinode.wit";
 pub const CACHE_DIR: &str = "/tmp/kinode-kit-cache";
 
 #[derive(Debug, Clone, Serialize, Deserialize)]
@@ -37,9 +33,7 @@ pub fn run_command(cmd: &mut Command) -> anyhow::Result<()> {
         Err(anyhow::anyhow!(
             "Command `{} {:?}` failed with exit code {}",
             cmd.get_program().to_str().unwrap(),
-            cmd.get_args()
-                .map(|a| a.to_str().unwrap())
-                .collect::<Vec<_>>(),
+            cmd.get_args().map(|a| a.to_str().unwrap()).collect::<Vec<_>>(),
             status.code().unwrap(),
         ))
     }
@@ -51,23 +45,23 @@ pub async fn download_file(url: &str, path: &Path) -> anyhow::Result<()> {
     let hex_url_path = format!("{}/{}", CACHE_DIR, hex_url);
     let hex_url_path = Path::new(&hex_url_path);
 
-    let content = if hex_url_path.exists() {
-        fs::read(hex_url_path).await?
-    } else {
-        let response = reqwest::get(url).await?;
+    let content =
+        if hex_url_path.exists() {
+            fs::read(hex_url_path).await?
+        } else {
+            let response = reqwest::get(url).await?;
 
-        // Check if response status is 200 (OK)
-        if response.status() != reqwest::StatusCode::OK {
-            return Err(anyhow::anyhow!(
-                "Failed to download file: HTTP Status {}",
-                response.status()
-            ));
-        }
+            // Check if response status is 200 (OK)
+            if response.status() != reqwest::StatusCode::OK {
+                return Err(anyhow::anyhow!("Failed to download file: HTTP Status {}", response.status()));
+            }
 
-        let content = response.bytes().await?.to_vec();
-        fs::write(hex_url_path, &content).await?;
-        content
-    };
+            let content = response.bytes()
+                .await?
+                .to_vec();
+            fs::write(hex_url_path, &content).await?;
+            content
+        };
 
     if path.exists() {
         if path.is_dir() {
@@ -79,61 +73,39 @@ pub async fn download_file(url: &str, path: &Path) -> anyhow::Result<()> {
             }
         }
     }
-    fs::create_dir_all(
-        path.parent()
-            .ok_or(anyhow::anyhow!("path doesn't have parent"))?,
-    )
-    .await?;
+    fs::create_dir_all(path.parent().ok_or(anyhow::anyhow!("path doesn't have parent"))?).await?;
     fs::write(path, &content).await?;
     Ok(())
 }
 
-async fn compile_javascript_wasm_process(process_dir: &Path, verbose: bool) -> anyhow::Result<()> {
-    println!(
-        "Compiling Javascript Kinode process in {:?}...",
-        process_dir
-    );
+async fn compile_javascript_wasm_process(
+    process_dir: &Path,
+    verbose: bool,
+) -> anyhow::Result<()> {
+    println!("Compiling Javascript Kinode process in {:?}...", process_dir);
     let wit_dir = process_dir.join("wit");
     download_file(KINODE_WIT_URL, &wit_dir.join("kinode.wit")).await?;
 
-    let wasm_file_name = process_dir.file_name().and_then(|s| s.to_str()).unwrap();
+    let wasm_file_name = process_dir
+        .file_name()
+        .and_then(|s| s.to_str())
+        .unwrap();
 
-    run_command(
-        Command::new("npm")
-            .arg("install")
-            .current_dir(process_dir)
-            .stdout(if verbose {
-                Stdio::inherit()
-            } else {
-                Stdio::null()
-            })
-            .stderr(if verbose {
-                Stdio::inherit()
-            } else {
-                Stdio::null()
-            }),
+    run_command(Command::new("npm")
+        .arg("install")
+        .current_dir(process_dir)
+        .stdout(if verbose { Stdio::inherit() } else { Stdio::null() })
+        .stderr(if verbose { Stdio::inherit() } else { Stdio::null() })
     )?;
 
-    run_command(
-        Command::new("node")
-            .args(&["componentize.mjs", wasm_file_name])
-            .current_dir(process_dir)
-            .stdout(if verbose {
-                Stdio::inherit()
-            } else {
-                Stdio::null()
-            })
-            .stderr(if verbose {
-                Stdio::inherit()
-            } else {
-                Stdio::null()
-            }),
+    run_command(Command::new("node")
+        .args(&["componentize.mjs", wasm_file_name])
+        .current_dir(process_dir)
+        .stdout(if verbose { Stdio::inherit() } else { Stdio::null() })
+        .stderr(if verbose { Stdio::inherit() } else { Stdio::null() })
     )?;
 
-    println!(
-        "Done compiling Javascript Kinode process in {:?}.",
-        process_dir
-    );
+    println!("Done compiling Javascript Kinode process in {:?}.", process_dir);
     Ok(())
 }
 
@@ -146,22 +118,16 @@ async fn compile_python_wasm_process(
     let wit_dir = process_dir.join("wit");
     download_file(KINODE_WIT_URL, &wit_dir.join("kinode.wit")).await?;
 
-    let wasm_file_name = process_dir.file_name().and_then(|s| s.to_str()).unwrap();
+    let wasm_file_name = process_dir
+        .file_name()
+        .and_then(|s| s.to_str())
+        .unwrap();
 
-    run_command(
-        Command::new(python)
-            .args(&["-m", "venv", PY_VENV_NAME])
-            .current_dir(process_dir)
-            .stdout(if verbose {
-                Stdio::inherit()
-            } else {
-                Stdio::null()
-            })
-            .stderr(if verbose {
-                Stdio::inherit()
-            } else {
-                Stdio::null()
-            }),
+    run_command(Command::new(python)
+        .args(&["-m", "venv", PY_VENV_NAME])
+        .current_dir(process_dir)
+        .stdout(if verbose { Stdio::inherit() } else { Stdio::null() })
+        .stderr(if verbose { Stdio::inherit() } else { Stdio::null() })
     )?;
     run_command(Command::new("bash")
         .args(&[
@@ -177,7 +143,10 @@ async fn compile_python_wasm_process(
     Ok(())
 }
 
-async fn compile_rust_wasm_process(process_dir: &Path, verbose: bool) -> anyhow::Result<()> {
+async fn compile_rust_wasm_process(
+    process_dir: &Path,
+    verbose: bool,
+) -> anyhow::Result<()> {
     println!("Compiling Rust Kinode process in {:?}...", process_dir);
 
     // Paths
@@ -195,7 +164,7 @@ async fn compile_rust_wasm_process(process_dir: &Path, verbose: bool) -> anyhow:
 
     // Check and download wasi_snapshot_preview1.wasm if it does not exist
     let wasi_snapshot_file = process_dir.join("wasi_snapshot_preview1.wasm");
-    let wasi_version = "15.0.1"; // TODO: un-hardcode
+    let wasi_version = "15.0.1";  // TODO: un-hardcode
     let wasi_snapshot_url = format!(
         "https://github.com/bytecodealliance/wasmtime/releases/download/v{}/wasi_snapshot_preview1.reactor.wasm",
         wasi_version,
@@ -203,130 +172,80 @@ async fn compile_rust_wasm_process(process_dir: &Path, verbose: bool) -> anyhow:
     download_file(&wasi_snapshot_url, &wasi_snapshot_file).await?;
 
     // Create target.wasm (compiled .wit) & world
-    run_command(
-        Command::new("wasm-tools")
-            .args(&[
-                "component",
-                "wit",
-                wit_dir.to_str().unwrap(),
-                "-o",
-                &bindings_dir.join("target.wasm").to_str().unwrap(),
-                "--wasm",
-            ])
-            .stdout(if verbose {
-                Stdio::inherit()
-            } else {
-                Stdio::null()
-            })
-            .stderr(if verbose {
-                Stdio::inherit()
-            } else {
-                Stdio::null()
-            }),
+    run_command(Command::new("wasm-tools")
+        .args(&["component", "wit",
+            wit_dir.to_str().unwrap(),
+            "-o", &bindings_dir.join("target.wasm").to_str().unwrap(),
+            "--wasm",
+        ])
+        .stdout(if verbose { Stdio::inherit() } else { Stdio::null() })
+        .stderr(if verbose { Stdio::inherit() } else { Stdio::null() })
     )?;
 
     // Copy wit directory to bindings
     fs::create_dir_all(&bindings_dir.join("wit")).await?;
     let mut entries = fs::read_dir(&wit_dir).await?;
     while let Some(entry) = entries.next_entry().await? {
-        fs::copy(
-            entry.path(),
-            bindings_dir.join("wit").join(entry.file_name()),
-        )
-        .await?;
+        fs::copy(entry.path(), bindings_dir.join("wit").join(entry.file_name())).await?;
     }
 
     // Create an empty world file
     File::create(bindings_dir.join("world"))?;
 
     // Build the module using Cargo
-    run_command(
-        Command::new("cargo")
-            .args(&[
-                "+nightly",
-                "build",
-                "--release",
-                "--no-default-features",
-                "--target",
-                "wasm32-wasi",
-                "--target-dir",
-                "target",
-            ])
-            .current_dir(process_dir)
-            .stdout(if verbose {
-                Stdio::inherit()
-            } else {
-                Stdio::null()
-            })
-            .stderr(if verbose {
-                Stdio::inherit()
-            } else {
-                Stdio::null()
-            }),
+    run_command(Command::new("cargo")
+        .args(&["+nightly", "build",
+            "--release",
+            "--no-default-features",
+            "--target", "wasm32-wasi",
+            "--target-dir", "target",
+        ])
+        .current_dir(process_dir)
+        .stdout(if verbose { Stdio::inherit() } else { Stdio::null() })
+        .stderr(if verbose { Stdio::inherit() } else { Stdio::null() })
     )?;
 
     // Adapt the module using wasm-tools
 
     // For use inside of process_dir
-    let wasm_file_name = process_dir.file_name().and_then(|s| s.to_str()).unwrap();
+    let wasm_file_name = process_dir
+        .file_name()
+        .and_then(|s| s.to_str())
+        .unwrap();
 
     let wasm_file_prefix = Path::new("target/wasm32-wasi/release");
-    let wasm_file = wasm_file_prefix.join(&format!("{}.wasm", wasm_file_name));
-    let adapted_wasm_file = wasm_file_prefix.join(&format!("{}_adapted.wasm", wasm_file_name));
+    let wasm_file = wasm_file_prefix
+        .join(&format!("{}.wasm", wasm_file_name));
+    let adapted_wasm_file = wasm_file_prefix
+        .join(&format!("{}_adapted.wasm", wasm_file_name));
 
     let wasi_snapshot_file = Path::new("wasi_snapshot_preview1.wasm");
 
-    run_command(
-        Command::new("wasm-tools")
-            .args(&[
-                "component",
-                "new",
-                wasm_file.to_str().unwrap(),
-                "-o",
-                adapted_wasm_file.to_str().unwrap(),
-                "--adapt",
-                wasi_snapshot_file.to_str().unwrap(),
-            ])
-            .current_dir(process_dir)
-            .stdout(if verbose {
-                Stdio::inherit()
-            } else {
-                Stdio::null()
-            })
-            .stderr(if verbose {
-                Stdio::inherit()
-            } else {
-                Stdio::null()
-            }),
+    run_command(Command::new("wasm-tools")
+        .args(&["component", "new",
+            wasm_file.to_str().unwrap(),
+            "-o", adapted_wasm_file.to_str().unwrap(),
+            "--adapt", wasi_snapshot_file.to_str().unwrap(),
+        ])
+        .current_dir(process_dir)
+        .stdout(if verbose { Stdio::inherit() } else { Stdio::null() })
+        .stderr(if verbose { Stdio::inherit() } else { Stdio::null() })
     )?;
 
     let wasm_path = format!("../pkg/{}.wasm", wasm_file_name);
     let wasm_path = Path::new(&wasm_path);
 
     // Embed wit into the component and place it in the expected location
-    run_command(
-        Command::new("wasm-tools")
-            .args(&[
-                "component",
-                "embed",
-                wit_dir.strip_prefix(process_dir).unwrap().to_str().unwrap(),
-                "--world",
-                "process",
-                adapted_wasm_file.to_str().unwrap(),
-                "-o",
-                wasm_path.to_str().unwrap(),
-            ])
-            .current_dir(process_dir)
-            .stdout(if verbose {
-                Stdio::inherit()
-            } else {
-                Stdio::null()
-            })
-            .stderr(if verbose {
-                Stdio::inherit()
-            } else {
-                Stdio::null()
-            }),
+    run_command(Command::new("wasm-tools")
+        .args(&["component", "embed",
+            wit_dir.strip_prefix(process_dir).unwrap().to_str().unwrap(),
+            "--world", "process",
+            adapted_wasm_file.to_str().unwrap(),
+            "-o", wasm_path.to_str().unwrap(),
+        ])
+        .current_dir(process_dir)
+        .stdout(if verbose { Stdio::inherit() } else { Stdio::null() })
+        .stderr(if verbose { Stdio::inherit() } else { Stdio::null() })
     )?;
 
     println!("Done compiling Rust Kinode process in {:?}.", process_dir);
@@ -341,60 +260,33 @@ async fn compile_and_copy_ui(package_dir: &Path, verbose: bool) -> anyhow::Resul
         if ui_path.join("package.json").exists() {
             println!("UI directory found, running npm install...");
 
-            run_command(
-                Command::new("bash")
-                    .arg("-c")
-                    .arg("source ~/.nvm/nvm.sh && npm install")
-                    .current_dir(&ui_path)
-                    .stdout(if verbose {
-                        Stdio::inherit()
-                    } else {
-                        Stdio::null()
-                    })
-                    .stderr(if verbose {
-                        Stdio::inherit()
-                    } else {
-                        Stdio::null()
-                    }),
+            run_command(Command::new("bash")
+                .arg("-c")
+                .arg("source ~/.nvm/nvm.sh && npm install")
+                .current_dir(&ui_path)
+                .stdout(if verbose { Stdio::inherit() } else { Stdio::null() })
+                .stderr(if verbose { Stdio::inherit() } else { Stdio::null() })
             )?;
 
             println!("Running npm run build:copy...");
 
-            run_command(
-                Command::new("bash")
-                    .arg("-c")
-                    .arg("source ~/.nvm/nvm.sh && npm run build:copy")
-                    .current_dir(&ui_path)
-                    .stdout(if verbose {
-                        Stdio::inherit()
-                    } else {
-                        Stdio::null()
-                    })
-                    .stderr(if verbose {
-                        Stdio::inherit()
-                    } else {
-                        Stdio::null()
-                    }),
+            run_command(Command::new("bash")
+                .arg("-c")
+                .arg("source ~/.nvm/nvm.sh && npm run build:copy")
+                .current_dir(&ui_path)
+                .stdout(if verbose { Stdio::inherit() } else { Stdio::null() })
+                .stderr(if verbose { Stdio::inherit() } else { Stdio::null() })
             )?;
         } else {
             let pkg_ui_path = package_dir.join("pkg/ui");
             if pkg_ui_path.exists() {
                 fs::remove_dir_all(&pkg_ui_path).await?;
             }
-            run_command(
-                Command::new("cp")
-                    .args(["-r", "ui", "pkg/ui"])
-                    .current_dir(&package_dir)
-                    .stdout(if verbose {
-                        Stdio::inherit()
-                    } else {
-                        Stdio::null()
-                    })
-                    .stderr(if verbose {
-                        Stdio::inherit()
-                    } else {
-                        Stdio::null()
-                    }),
+            run_command(Command::new("cp")
+                .args(["-r", "ui", "pkg/ui"])
+                .current_dir(&package_dir)
+                .stdout(if verbose { Stdio::inherit() } else { Stdio::null() })
+                .stderr(if verbose { Stdio::inherit() } else { Stdio::null() })
             )?;
         }
     } else {
@@ -434,17 +326,19 @@ async fn compile_package(
                 let python = check_py_deps()?;
                 compile_python_wasm_process(&path, &python, verbose).await?;
             } else if path.join(JAVASCRIPT_SRC_PATH).exists() {
-                let _old_node_version = if skip_deps_check {
-                    OldNodeVersion::none()
-                } else {
-                    let (deps, old_node_version) = check_js_deps()?;
-                    if deps.is_empty() {
-                        old_node_version
+                let _old_node_version =
+                    if skip_deps_check {
+                        OldNodeVersion::none()
                     } else {
-                        get_deps(deps)?;
-                        set_newest_valid_node_version(None, None)?.unwrap_or(OldNodeVersion::none())
-                    }
-                };
+                        let (deps, old_node_version) = check_js_deps()?;
+                        if deps.is_empty() {
+                            old_node_version
+                        } else {
+                            get_deps(deps)?;
+                            set_newest_valid_node_version(None, None)?
+                                .unwrap_or(OldNodeVersion::none())
+                        }
+                    };
                 compile_javascript_wasm_process(&path, verbose).await?;
             }
         }
@@ -470,9 +364,7 @@ pub async fn execute(
     let ui_dir = package_dir.join("ui");
     if !ui_dir.exists() {
         if ui_only {
-            return Err(anyhow::anyhow!(
-                "kit build: can't build UI: no ui directory exists"
-            ));
+            return Err(anyhow::anyhow!("kit build: can't build UI: no ui directory exists"));
         } else {
             compile_package(package_dir, verbose, skip_deps_check).await
         }
@@ -481,13 +373,14 @@ pub async fn execute(
             return compile_package(package_dir, verbose, skip_deps_check).await;
         }
 
-        let _old_node_version = if skip_deps_check {
-            OldNodeVersion::none()
-        } else {
-            let (deps, old_node_version) = check_js_deps()?;
-            get_deps(deps)?;
-            old_node_version
-        };
+        let _old_node_version =
+            if skip_deps_check {
+                OldNodeVersion::none()
+            } else {
+                let (deps, old_node_version) = check_js_deps()?;
+                get_deps(deps)?;
+                old_node_version
+            };
 
         if ui_only {
             compile_and_copy_ui(package_dir, verbose).await
